move splitTermIntoSegements into its own file
